Restructured VS proj namespace/folders
As per the discussion on #4, I've grouped the data-resolvers into folders by their type (Property or GridCell resolvers).
This will reduce the number of namespaces used in the project/assembly.
