+++ src/Umbraco.Courier.Contrib.Resolvers/GridCellDataResolvers/DocTypeGridEditorGridCellResolver.cs
@@ -10,7 +10,7 @@ using Umbraco.Courier.Core.ProviderModel;
 using Umbraco.Courier.DataResolvers.PropertyDataResolvers;
 using Umbraco.Courier.ItemProviders;
 
-namespace Umbraco.Courier.Contrib.Resolvers.DocTypeGridEditor
+namespace Umbraco.Courier.Contrib.Resolvers.GridCellDataResolvers
 {
     /// <summary>
     /// DocTypeGridEditor Grid Cell Resolver for DTGE by Matt Brailsford & Lee Kelleher.
